backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.10.1
pydantic-settings==2.6.1
python-multipart==0.0.12
websockets==12.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
httpx==0.27.2
python-dotenv==1.0.1
PyJWT==2.9.0
PyYAML==6.0.2